Separate DefaultDeviceAttestationVerifier from credentials
- This PR is on the way to resolving #15209.
- This PR does the following:
  - Splits DefaultDeviceAttesationVerifier from the main src/credentials target
    since it is an optional component that can be overridden by different
    commissioners
  - Adds the beginning of plumbing to properly select the Trust Store for
    the DefaultDeviceAttestationVerifier.
  - Moved DefaultDeviceAttestatationVerifier from credentials/examples to
    credentials/attestation_verifier

Missing, to come in the follow-up:
- The implementation of a new file-based PAA trust store configured with
  the path passed in the plumbing added here.

Testing done: unit tests and cert tests still pass. Commissioning still works
