+++ examples/chip-tool-darwin/commands/common/CHIPCommandBridge.h
@@ -62,7 +62,7 @@ protected:
     CHIPDeviceController * CurrentCommissioner();
 
 private:
-    CHIP_ERROR InitializeCommissioner(std::string key, chip::FabricId fabricId);
+    CHIP_ERROR InitializeCommissioner(std::string key, chip::FabricId fabricId, const chip::Credentials::AttestationTrustStore trustStore *);
     CHIP_ERROR ShutdownCommissioner();
     uint16_t CurrentCommissionerIndex();
