+++ examples/platform/linux/BUILD.gn
@@ -45,6 +45,7 @@ source_set("app-main") {
 
   public_deps = [
     "${chip_root}/src/app/server",
+    "${chip_root}/src/credentials:default_attestation_verifier",
     "${chip_root}/src/lib",
     "${chip_root}/src/lib/shell",
     "${chip_root}/src/lib/shell:shell_core",
