+++ src/credentials/BUILD.gn
@@ -32,15 +32,15 @@ static_library("credentials") {
     "DeviceAttestationCredsProvider.cpp",
     "DeviceAttestationCredsProvider.h",
     "DeviceAttestationVendorReserved.h",
-    "DeviceAttestationVerifier.cpp",
-    "DeviceAttestationVerifier.h",
     "FabricTable.cpp",
     "FabricTable.h",
     "GenerateChipX509Cert.cpp",
     "GroupDataProvider.h",
     "GroupDataProviderImpl.cpp",
-    "examples/DefaultDeviceAttestationVerifier.cpp",
-    "examples/DefaultDeviceAttestationVerifier.h",
+
+    "attestation_verifier/DeviceAttestationVerifier.cpp",
+    "attestation_verifier/DeviceAttestationVerifier.h",
+
     "examples/DeviceAttestationCredsExample.cpp",
     "examples/DeviceAttestationCredsExample.h",
     "examples/ExampleDACs.cpp",
@@ -73,10 +73,29 @@ static_library("credentials") {
 
   public_deps = [
     "${chip_root}/src/crypto",
+    "${chip_root}/src/platform",
     "${chip_root}/src/lib/asn1",
     "${chip_root}/src/lib/core",
     "${chip_root}/src/lib/support",
-    "${chip_root}/src/transport/raw",
+    "${nlassert_root}:nlassert",
+  ]
+}
+
+static_library("default_attestation_verifier") {
+  output_name = "libDefaultAttestationVerifier"
+
+  sources = [
+    "attestation_verifier/DefaultDeviceAttestationVerifier.cpp",
+    "attestation_verifier/DefaultDeviceAttestationVerifier.h",
+  ]
+
+  if (chip_device_platform == "esp32" || chip_device_platform == "nrfconnect") {
+    defines = [ "CURRENT_TIME_NOT_IMPLEMENTED=1" ]
+  }
+
+  public_deps = [
+    ":credentials",
+    "${chip_root}/src/crypto",
     "${nlassert_root}:nlassert",
   ]
 }
