+++ src/credentials/tests/BUILD.gn
@@ -19,7 +19,7 @@ import("//build_overrides/nlunit_test.gni")
 import("${chip_root}/build/chip/chip_test_suite.gni")
 
 static_library("cert_test_vectors") {
-  output_name = "libChipCertTestVectors"
+  output_name = "libCertTestVectors"
   output_dir = "${root_out_dir}/lib"
 
   sources = [
@@ -35,7 +35,7 @@ static_library("cert_test_vectors") {
 }
 
 chip_test_suite("tests") {
-  output_name = "libChipCredentials"
+  output_name = "libCredentialsTest"
   output_dir = "${root_out_dir}/lib"
 
   test_sources = [
@@ -52,6 +52,7 @@ chip_test_suite("tests") {
   public_deps = [
     ":cert_test_vectors",
     "${chip_root}/src/credentials",
+    "${chip_root}/src/credentials:default_attestation_verifier",
     "${chip_root}/src/lib/core",
     "${nlunit_test_root}:nlunit-test",
   ]
