+++ src/darwin/Framework/CHIP/BUILD.gn
@@ -68,6 +68,7 @@ static_library("framework") {
   public_deps = [
     "${chip_root}/src/controller",
     "${chip_root}/src/controller/data_model",
+    "${chip_root}/src/credentials:default_attestation_verifier",
     "${chip_root}/src/lib/core",
     "${chip_root}/src/lib/support",
   ]
